+++ uniCloud-aliyun/cloudfunctions/router/service/admin/system_uni/components-dynamic/sys/delete.js
@@ -21,7 +21,7 @@ module.exports = {
 		}
 		// 执行数据库API请求
 		let num = await vk.baseDao.del({
-		dbName:"opendb-components-dynamic",
+		dbName:"vk-components-dynamic",
 			whereJson:{
 				_id
 			}