+++ uniCloud-aliyun/cloudfunctions/router/service/admin/system_uni/uni-id-files/categories/sys/update.js
@@ -28,7 +28,7 @@ module.exports = {
 		if(vk.pubfn.isNull(name)){
 			return { code : -1, msg : '名称不能为空' };
 		}
-		let dbName = "uni-id-files-categories";
+		let dbName = "vk-files-categories";
 
 		// 检测data_id是否已存在
 		let num = await vk.baseDao.count({