+++ uniCloud-aliyun/cloudfunctions/router/service/admin/system_uni/uni-id-files/files/kh/update.js
@@ -27,7 +27,7 @@ module.exports = {
 		if (vk.pubfn.isNullOne(_id)) {
 			return { code: -1, msg: '_id不能为空' };
 		}
-		let dbName = "uni-id-files";
+		let dbName = "vk-files";
 		let whereJson = {};
 		let dataJson = {
 			display_name,